+++ requirements.txt
@@ -1,0 +1,4 @@
+python-dateutil~=2.8.1
+pytz~=2020.1
+requests~=2.24.0
+pydash~=4.8.0
